TRIPNEST_LLM/tripnest_llm/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
google-generativeai>=0.8.5
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=1.10.0
faiss-cpu>=1.7.4
httpx>=0.24.0